Django-Strawberry-GraphQL/requirements.txt
django
strawberry-graphql
strawberry-graphql-django
gql[all]
psycopg2